Fixed issues for code review.
